% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{combinations}
\alias{combinations}
\title{Combinations}
\usage{
combinations(
  n,
  r,
  v = 1:n,
  set = TRUE,
  repeats.allowed = FALSE,
  count.permutations = FALSE,
  out_format = "auto"
)
}
\arguments{
\item{n}{Numeric. Number of elements in the vector.}

\item{r}{Numeric. Number of elements to choose.}

\item{v}{vector or list containing the objects that should be combined. Default sequence from 1 to n.}

\item{set}{Logical. If TRUE, remove duplicate values from the input vector.}

\item{repeats.allowed}{Logical. If TRUE, combinations with repetition are allowed.}

\item{count.permutations}{Logical. If TRUE, returns the number of permutations corresponding to each combination.}

\item{out_format}{Character. The desired format for the output, either "matrix", "dataframe", "tibble", "datatable", or "auto".}
}
\value{
A matrix, dataframe, tibble, or datatable with the computed combinations.
}
\description{
Compute combinations of set of objects
}
\examples{
combinations(n = 4, r = 2)
combinations(n = 4, r = 2, repeats.allowed = TRUE)
combinations(n = 4, r = 2, repeats.allowed = TRUE, count.permutations = TRUE)
}
