% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permutations}
\alias{permutations}
\title{Permutations}
\usage{
permutations(
  n,
  r,
  v = 1:n,
  set = TRUE,
  repeats.allowed = FALSE,
  count.permutations = FALSE,
  out_format = "auto"
)
}
\arguments{
\item{n}{Numeric. Number of elements in the vector.}

\item{r}{Numeric. Number of elements to permute.}

\item{v}{vector or list containing the objects that should be permuted. Default sequence from 1 to n.}

\item{set}{Logical. If TRUE, remove duplicate values from the input vector.}

\item{repeats.allowed}{Logical. If TRUE, permutations with repetition are allowed.}

\item{count.permutations}{Logical. If TRUE, returns the number of identical permutations for each permutation.}

\item{out_format}{Character. The desired format for the output, either "matrix", "dataframe", "tibble", "datatable", or "auto".}
}
\value{
A matrix, dataframe, tibble, or datatable with the computed permutations.
}
\description{
Compute permutations of a set of objects.
}
\examples{
permutations(n = 4, r = 2)
permutations(n = 4, r = 2, repeats.allowed = TRUE)
permutations(n = 4, r = 2, repeats.allowed = TRUE, count.permutations = TRUE)
}
