% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{combinations}
\alias{combinations}
\title{Combinations}
\usage{
combinations(
  n,
  r,
  v = 1:n,
  set = TRUE,
  repeats.allowed = FALSE,
  count.permutations = FALSE,
  out_format = "auto"
)
}
\arguments{
\item{n}{Numeric. Number of elements in the vector.}

\item{r}{Numeric. Number of elements to choose.}

\item{v}{Numeric vector. Default sequence from 1 to n.}

\item{set}{Logical. If TRUE, remove duplicate values from the input vector.}

\item{repeats.allowed}{Logical. If TRUE, combinations with repetition are allowed.}

\item{count.permutations}{Logical. If TRUE, returns the number of permutations corresponding to each combination.}

\item{out_format}{Character. Either 'auto' or another desired format. 'auto' will try to convert to a data.table.}
}
\value{
A data.table or matrix with the computed combinations.
}
\description{
Compute combinations from a vector.
}
\examples{
combinations(n = 4, r = 2)
combinations(n = 4, r = 2, repeats.allowed = TRUE)
combinations(n = 4, r = 2, count.permutations = TRUE)
}
